% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmif.cv.R
\name{mmif.cv}
\alias{mmif.cv}
\title{The Cross Validation of Random Forest}
\usage{
mmif.cv(target.mat, feature.mat, param_grid, n_folds = 3, scorer = NULL,
  n_cpu = 1)
}
\arguments{
\item{target.mat}{The response variable of the model}

\item{feature.mat}{a data frame containing the feature variables in the model.}

\item{param_grid}{the list of paramaters}

\item{n_folds}{The number of folds}

\item{scorer}{The Loss calculation function}

\item{n_cpu}{The number of cores to register for parallel programing of the code, default value is 1 and n_cpu = -1 to select all cores.}
}
\value{
The list consist of best score, best parameters and list of all parameter values with cross validation score .
}
\description{
Performing grid search to select the best parameters via cross validation on the random forest.
}
\examples{
library(mmit)
target.mat <- rbind(
  c(0,1), c(0,1), c(0,1),
  c(2,3), c(2,3), c(2,3))

feature.mat <- rbind(
  c(1,0,0), c(1,1,0), c(1,2,0),
  c(1,3,0), c(1,4,0), c(1,5,0))

colnames(feature.mat) <- c("a", "b", "c")
feature.mat <- data.frame(feature.mat)

param_grid <- NULL
param_grid$max_depth <- c(Inf, 4, 3)
param_grid$margin <- c(2, 3, 5)
param_grid$min_sample <- c(2, 5, 10)
param_grid$loss <- c("hinge")
param_grid$n_trees <- c(10, 20, 30)
param_grid$n_features <- c(as.integer(ncol(feature.mat)**0.5))

result <- mmif.cv(target.mat, feature.mat, param_grid, scorer = mse, n_cpu = -1)

}
\author{
Toby Dylan Hocking, Alexandre Drouin, Torsten Hothorn, Parismita Das
}
