% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmif.predict.R
\name{mmif.predict}
\alias{mmif.predict}
\title{Prediction for Random Forest}
\usage{
mmif.predict(forest, test_feature.mat = NULL)
}
\arguments{
\item{forest}{Ensembe of forest}

\item{test_feature.mat}{a data frame containing the test feature variables in the model.}
}
\value{
Prediction values
}
\description{
Predictions as average of all predictions of the random forest.
}
\examples{
library(mmit)
target.mat <- rbind(
  c(0,1), c(0,1), c(0,1),
  c(2,3), c(2,3), c(2,3))

feature.mat <- rbind(
  c(1,0,0), c(1,1,0), c(1,2,0),
  c(1,3,0), c(1,4,0), c(1,5,0))

colnames(feature.mat) <- c("a", "b", "c")
feature.mat <- data.frame(feature.mat)

forest <- mmif(target.mat, feature.mat)
pred <- mmif.predict(forest, feature.mat)

}
\author{
Toby Dylan Hocking, Alexandre Drouin, Torsten Hothorn, Parismita Das
}
