% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmit.pruning.R
\name{mmit.pruning}
\alias{mmit.pruning}
\title{The Pruned Max Margin Interval Tree}
\usage{
mmit.pruning(tree)
}
\arguments{
\item{tree}{The fitted tree using \code{"mmit()"} function}
}
\value{
The learned regression tree as an object of class party.
}
\description{
Pruning the regression tree for censored data.
}
\examples{
library(mmit)
target.mat <- rbind(
  c(0,1), c(0,1), c(0,1),
  c(2,3), c(2,3), c(2,3))

feature.mat <- rbind(
  c(1,0,0), c(1,1,0), c(1,2,0),
  c(1,3,0), c(1,4,0), c(1,5,0))

colnames(feature.mat) <- c("a", "b", "c")
feature.mat <- data.frame(feature.mat)


tree <- mmit(feature.mat, target.mat)

pruned_tree <- mmit.pruning(tree)

}
\author{
Toby Dylan Hocking, Alexandre Drouin, Torsten Hothorn, Parismita Das
}
