% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{mse}
\alias{mse}
\title{The Mean Square Error}
\usage{
mse(y_true, y_pred)
}
\arguments{
\item{y_true}{The actual response variable of the model}

\item{y_pred}{The predicted response value of the model}
}
\value{
A numeric value which signifies the error quantity.
}
\description{
Metric for mean aquare error calculation.
}
\examples{
library(mmit)
y_true <- rbind(
  c(0,1), c(0,1), c(0,1),
  c(2,3), c(2,3), c(2,3))

y_pred <- c(0.5, 2, 0, 1.5, 3.5, 2.5)

out <- mse(y_true, y_pred)

}
\author{
Toby Dylan Hocking, Alexandre Drouin, Torsten Hothorn, Parismita Das
}
