% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmif.cv.predict.R
\name{predict.mmif.cv}
\alias{predict.mmif.cv}
\title{The Predict Function for Cross Validation of Random Forest}
\usage{
\method{predict}{mmif.cv}(object, newdata = NULL, perm = NULL)
}
\arguments{
\item{object}{Object obtained from \code{"mmif.cv()"}}

\item{newdata}{an optional data frame containing the testing data which is to be predicted.}

\item{perm}{an optional character vector of variable names.}
}
\value{
Predictions Average output of each tree in the forest
}
\description{
Fits the new data into the mmif.cv model to give prediction values
}
\examples{
library(mmit)
target.mat <- rbind(
  c(0,1), c(0,1), c(0,1),
  c(2,3), c(2,3), c(2,3))

feature.mat <- rbind(
  c(1,0,0), c(1,1,0), c(1,2,0),
  c(1,3,0), c(1,4,0), c(1,5,0))

colnames(feature.mat) <- c("a", "b", "c")
feature.mat <- data.frame(feature.mat)

fit <- mmif.cv(feature.mat, target.mat)

pred <- predict(fit)

}
\author{
Toby Dylan Hocking, Alexandre Drouin, Torsten Hothorn, Parismita Das
}
