% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmit.predict.R
\name{predict.mmit}
\alias{predict.mmit}
\title{The Predict Function for Max Margin Interval Tree}
\usage{
\method{predict}{mmit}(object, newdata = NULL, perm = NULL, ...)
}
\arguments{
\item{object}{The Max Margin Interval Tree obtained from \code{"mmit()"}}

\item{newdata}{an optional data frame containing the testing data which is to be predicted.}

\item{perm}{an optional character vector of variable names.}

\item{\dots}{...}
}
\value{
The learned regression tree as an object of class party.
}
\description{
Fits the new data into the MMIT model to give prediction values
}
\examples{
library(mmit)
target.mat <- rbind(
  c(0,1), c(0,1), c(0,1),
  c(2,3), c(2,3), c(2,3))

feature.mat <- rbind(
  c(1,0,0), c(1,1,0), c(1,2,0),
  c(1,3,0), c(1,4,0), c(1,5,0))

colnames(feature.mat) <- c("a", "b", "c")
feature.mat <- data.frame(feature.mat)

tree <- mmit(feature.mat, target.mat)

pred <- predict.mmit(tree)

}
\author{
Toby Dylan Hocking, Alexandre Drouin, Torsten Hothorn, Parismita Das
}
