% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{zero_one_loss}
\alias{zero_one_loss}
\title{The Zero One Loss}
\usage{
zero_one_loss(y_true, y_pred)
}
\arguments{
\item{y_true}{The actual response variable of the model}

\item{y_pred}{The predicted response value of the model}
}
\value{
A numeric value which signifies the error quantity.
}
\description{
Metric for error calculation where the function gives zero value inside the interval else one.
}
\examples{
library(mmit)
y_true <- rbind(
  c(0,1), c(0,1), c(0,1),
  c(2,3), c(2,3), c(2,3))

y_pred <- c(0.5, 2, 0, 1.5, 3.5, 2.5)

out <- zero_one_loss(y_true, y_pred)

}
\author{
Toby Dylan Hocking, Alexandre Drouin, Torsten Hothorn, Parismita Das
}
