% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmif.cv.R
\name{mmif.cv}
\alias{mmif.cv}
\title{Cross-validation for model selection with Random Forests of Max Margin Interval Trees}
\usage{
mmif.cv(target.mat, feature.mat, param_grid, n_folds = 3,
  scorer = NULL, future.seed = FALSE)
}
\arguments{
\item{target.mat}{The response variable of the model}

\item{feature.mat}{A data frame containing the feature variables in the model.}

\item{param_grid}{A list with values to try for each hyperparameter (max_depth, margin, min_sample, loss, n_trees, n_features).}

\item{n_folds}{The number of folds for k-fold cross-validation}

\item{scorer}{The function used to calculate the cross-validation score (e.g., mse, zero_one_loss)}

\item{future.seed}{A logical or an integer (of length one or seven), or a list of length(X) with pre-generated random seeds.}
}
\value{
The best score, best model (trained with best parameters), best parameters, and list of all parameter values with cross validation score.
}
\description{
Performing grid search to select the best hyperparameters of mmif via cross-validation.
}
\examples{
library(mmit)

target.mat <- rbind(
  c(0,1), c(0,1), c(0,1),
  c(2,3), c(2,3), c(2,3))

feature.mat <- rbind(
  c(1,0,0), c(1,1,0), c(1,2,0),
  c(1,3,0), c(1,4,0), c(1,5,0))

colnames(feature.mat) <- c("a", "b", "c")
feature.mat <- data.frame(feature.mat)

param_grid <- NULL
param_grid$max_depth <- c(Inf, 4, 3)
param_grid$margin <- c(2, 3, 5)
param_grid$min_sample <- c(2, 5, 10)
param_grid$loss <- c("hinge")
param_grid$n_trees <- c(10, 20, 30)
param_grid$n_features <- c(ceiling(ncol(feature.mat)**0.5))

set.seed(1)
result <- mmif.cv(target.mat, feature.mat, param_grid, scorer = mse, future.seed = TRUE)

}
\author{
Toby Dylan Hocking, Alexandre Drouin, Torsten Hothorn, Parismita Das
}
