% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmif.R
\name{mmif}
\alias{mmif}
\title{Random Forest of Max Margin Interval Tree}
\usage{
mmif(target.mat, feature.mat, max_depth = Inf, margin = 0, loss = "hinge",
  min_sample = 1, n_trees = 10,
  n_features = as.integer(ncol(feature.mat)^0.5), n_cpu = 1)
}
\arguments{
\item{target.mat}{The response variable of the model}

\item{feature.mat}{a data frame containing the feature variables in the model.}

\item{max_depth}{The maximum depth criteia}

\item{margin}{margin paramaters}

\item{loss}{The type of loss; (\code{"hinge"}, \code{"square"})}

\item{min_sample}{The minimum number of sample required}

\item{n_trees}{The Number of trees}

\item{n_features}{The number of features to be sampled.}

\item{n_cpu}{The number of cores to register for parallel programing of the code, default value is 1 and n_cpu = -1 to select all cores.}
}
\value{
List of ensemble of trees.
}
\description{
Performing random forest on ensemble of Max Margin Interval Tree.
}
\examples{
library(mmit)
target.mat <- rbind(
  c(0,1), c(0,1), c(0,1),
  c(2,3), c(2,3), c(2,3))

feature.mat <- rbind(
  c(1,0,0), c(1,1,0), c(1,2,0),
  c(1,3,0), c(1,4,0), c(1,5,0))

colnames(feature.mat) <- c("a", "b", "c")
feature.mat <- data.frame(feature.mat)

trees <- mmif(target.mat, feature.mat, margin = 2.0, n_cpu = -1)

}
\author{
Toby Dylan Hocking, Alexandre Drouin, Torsten Hothorn, Parismita Das
}
