% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmif.R
\name{mmif}
\alias{mmif}
\title{Random Forest of Max Margin Interval Tree}
\usage{
mmif(target.mat, feature.mat, test_feature.mat = feature.mat, parameters,
  n_trees = 10, n_example = 100,
  n_features = as.integer(ncol(feature.mat)^0.5))
}
\arguments{
\item{target.mat}{The response variable of the model}

\item{feature.mat}{a data frame containing the feature variables in the model.}

\item{test_feature.mat}{The testing data table.}

\item{parameters}{the list of paramaters such as max_depth, margin, loss, min_sample}

\item{n_trees}{The Number of trees}

\item{n_example}{The number of data elements to be sampled from dataset.}

\item{n_features}{The number of features to be sampled.}
}
\value{
predicted values of random forest
}
\description{
Performing grid search to select the best parameters via cross validation on the  a regression tree for censored data.
}
\examples{
library(mmit)
target.mat <- rbind(
  c(0,1), c(0,1), c(0,1),
  c(2,3), c(2,3), c(2,3))

feature.mat <- rbind(
  c(1,0,0), c(1,1,0), c(1,2,0),
  c(1,3,0), c(1,4,0), c(1,5,0))

colnames(feature.mat) <- c("a", "b", "c")
feature.mat <- data.frame(feature.mat)

testx <- data.frame(neuroblastomaProcessed$feature.mat)[45:90,]
testy <- neuroblastomaProcessed$target.mat[45:90,]
parameters <- list(max_depth = Inf, margin = 2.0, loss = "hinge", min_sample = 1)

pred <- mmif(target.mat, feature.mat, test, parameters = parameters)

}
\author{
Toby Dylan Hocking, Alexandre Drouin, Torsten Hothorn, Parismita Das
}
